# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ISM Tecnologia Financial System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relatorios/fechamento-mensal:
    get:
      operationId: getRelatorioFechamentoMensal
      tags: [relatorios]
      parameters:
        - name: mes
          in: query
          required: true
          schema:
            type: integer
        - name: ano
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly closing report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FechamentoMensalReport"
  /relatorios/dre:
    get:
      operationId: getRelatorioDre
      tags: [relatorios]
      parameters:
        - name: ano
          in: query
          required: true
          schema:
            type: integer
        - name: regime
          in: query
          schema:
            type: string
            enum: [competencia, caixa]
            default: competencia
      responses:
        "200":
          description: DRE report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DreReport"
  /relatorios/fluxo-caixa:
    get:
      operationId: getRelatorioFluxoCaixa
      tags: [relatorios]
      parameters:
        - name: ano
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cash flow report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FluxoCaixaReport"
  /relatorios/metas:
    get:
      operationId: getRelatorioMetas
      tags: [relatorios]
      parameters:
        - name: ano
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goals vs actual report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetasReport"
  /relatorios/contabil-fiscal:
    get:
      operationId: getRelatorioContabilFiscal
      tags: [relatorios]
      parameters:
        - name: data_inicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: filial_id
          in: query
          schema:
            type: integer
        - name: conta_id
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
            enum: [CP, CR, ambos]
            default: ambos
      responses:
        "200":
          description: Accounting/fiscal report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContabilFiscalItem"
...
  /tokens-api:
    get:
      operationId: getTokensApi
      tags: [tokens-api]
      responses:
        "200":
          description: List of API tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenApi"
    post:
      operationId: createTokenApi
      tags: [tokens-api]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenApi"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenApi"
  /tokens-api/{id}:
    put:
      operationId: updateTokenApi
      tags: [tokens-api]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTokenApi"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenApi"
    delete:
      operationId: deleteTokenApi
      tags: [tokens-api]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    FechamentoMensalReport:
      type: object
      properties:
        mes:
          type: integer
        ano:
          type: integer
        planejado_receber:
          type: number
        realizado_receber:
          type: number
        planejado_gastar:
          type: number
        realizado_gastar:
          type: number
        recebimentos:
          type: array
          items:
            $ref: "#/components/schemas/FechamentoCategoria"
        despesas:
          type: array
          items:
            $ref: "#/components/schemas/FechamentoCategoria"
      required:
        - mes
        - ano
        - planejado_receber
        - realizado_receber
        - planejado_gastar
        - realizado_gastar
        - recebimentos
        - despesas
    FechamentoCategoria:
      type: object
      properties:
        categoria:
          type: string
        planejado:
          type: number
        realizado:
          type: number
        percentual:
          type: number
      required:
        - categoria
        - planejado
        - realizado
        - percentual
    DreReport:
      type: object
      properties:
        ano:
          type: integer
        regime:
          type: string
        linhas:
          type: array
          items:
            $ref: "#/components/schemas/DreLinhaReport"
      required:
        - ano
        - regime
        - linhas
    DreLinhaReport:
      type: object
      properties:
        descricao:
          type: string
        tipo:
          type: string
          enum: [header, categoria, total, subtotal]
        valores:
          type: object
          additionalProperties:
            type: object
            properties:
              valor:
                type: number
              percentual:
                type: number
      required:
        - descricao
        - tipo
        - valores
    FluxoCaixaReport:
      type: object
      properties:
        ano:
          type: integer
        secoes:
          type: array
          items:
            $ref: "#/components/schemas/FluxoCaixaSecao"
      required:
        - ano
        - secoes
    FluxoCaixaSecao:
      type: object
      properties:
        titulo:
          type: string
        tipo:
          type: string
          enum: [saldo_inicial, entradas, saidas, transferencias, total, saldo_final]
        linhas:
          type: array
          items:
            $ref: "#/components/schemas/FluxoCaixaLinha"
      required:
        - titulo
        - tipo
        - linhas
    FluxoCaixaLinha:
      type: object
      properties:
        descricao:
          type: string
        codigo:
          type: string
        valores:
          type: object
          additionalProperties:
            type: number
      required:
        - descricao
        - valores
    MetasReport:
      type: object
      properties:
        ano:
          type: integer
        recebimentos:
          type: array
          items:
            $ref: "#/components/schemas/MetasLinhaReport"
        despesas:
          type: array
          items:
            $ref: "#/components/schemas/MetasLinhaReport"
        resultado:
          type: array
          items:
            $ref: "#/components/schemas/MetasLinhaReport"
      required:
        - ano
        - recebimentos
        - despesas
        - resultado
    MetasLinhaReport:
      type: object
      properties:
        categoria:
          type: string
        valores:
          type: object
          additionalProperties:
            type: object
            properties:
              previsto:
                type: number
              realizado:
                type: number
      required:
        - categoria
        - valores
    ContabilFiscalItem:
      type: object
      properties:
        conta_bancaria:
          type: string
        data:
          type: string
        descricao:
          type: string
        nome_parceiro:
          type: string
        valor:
          type: number
        categoria:
          type: string
        tipo:
          type: string
          enum: [CP, CR]
      required:
        - conta_bancaria
        - data
        - descricao
        - valor
        - tipo
...
    TokenApi:
      type: object
      properties:
        id:
          type: integer
        descricao:
          type: string
        token:
          type: string
        data_expiracao:
          type: string
          format: date
        ativo:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - descricao
        - ativo
    CreateTokenApi:
      type: object
      properties:
        descricao:
          type: string
        data_expiracao:
          type: string
          format: date
        ativo:
          type: boolean
      required:
        - descricao
    UpdateTokenApi:
      type: object
      properties:
        descricao:
          type: string
        data_expiracao:
          type: string
          format: date
        ativo:
          type: boolean
...