# lib/api-zod/src/generated/types/tokenApi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISM Tecnologia Financial System API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenApi {
  id: number;
  descricao: string;
  token?: string;
  data_expiracao?: Date;
  ativo: boolean;
  created_at?: string;
}