# lib/api-zod/src/generated/types/updateTokenApi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISM Tecnologia Financial System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTokenApi {
  descricao?: string;
  data_expiracao?: Date;
  ativo?: boolean;
}